Update tweets api route and remove stamps
